Organise the code, add error passing instead of unwraps
