+++ node-wasm/Cargo.toml
@@ -36,6 +36,7 @@ js-sys = "0.3.64"
 serde = { version = "1.0.164", features = ["derive"] }
 serde-wasm-bindgen = "0.6.0"
 serde_repr = "0.1"
+thiserror = "1.0"
 time = { version = "0.3", features = ["wasm-bindgen"] }
 tokio =  { version = "*", features = ["sync"]}
 tracing = "0.1.37"
