+++ node-wasm/src/node.rs
@@ -16,15 +16,17 @@ use lumina_node::node::NodeConfig;
 use lumina_node::store::IndexedDbStore;
 
 use crate::utils::{js_value_from_display, JsContext, Network};
-use crate::worker::SharedWorkerChannel;
-use crate::worker::{
+use crate::worker::commands::{
     GetConnectedPeers, GetHeader, GetListeners, GetLocalPeerId, GetMultipleHeaders, GetNetworkInfo,
     GetPeerTrackerInfo, GetSamplingMetadata, GetSyncerInfo, IsRunning, LastSeenNetworkHead,
     MultipleHeaderQuery, NodeCommand, NodeResponse, RequestHeader, RequestMultipleHeaders,
     SetPeerTrust, SingleHeaderQuery, StartNode, WaitConnected,
 };
+use crate::worker::SharedWorkerChannel;
 use crate::wrapper::libp2p::NetworkInfoSnapshot;
 use crate::Result;
+
+const LUMINA_SHARED_WORKER_NAME: &str = "lumina";
 
 /// Config for the lumina wasm node.
 #[wasm_bindgen(js_name = NodeConfig)]
@@ -40,7 +42,6 @@ pub struct WasmNodeConfig {
     pub bootnodes: Vec<String>,
 }
 
-// TODO: add on_error handler
 #[wasm_bindgen]
 struct NodeDriver {
     _worker: SharedWorker,
@@ -50,11 +51,14 @@ struct NodeDriver {
 
 #[wasm_bindgen]
 impl NodeDriver {
+    /// Create a new connection to a Lumina node in a Shared Worker.
+    /// Note that single Shared Worker can be accessed from multiple tabs, so Lumina may already
+    /// be running be running, before `NodeDriver::start` call.
     #[wasm_bindgen(constructor)]
     pub async fn new() -> NodeDriver {
         let mut opts = WorkerOptions::new();
         opts.type_(WorkerType::Module);
-        opts.name("lumina");
+        opts.name(LUMINA_SHARED_WORKER_NAME);
         let worker = SharedWorker::new_with_worker_options("/js/worker.js", &opts)
             .expect("could not worker");
 
@@ -72,104 +76,111 @@ impl NodeDriver {
         }
     }
 
-    pub async fn is_running(&mut self) -> bool {
-        let response = self.channel.send(IsRunning);
+    /// Check whether Lumina is currently running
+    pub async fn is_running(&mut self) -> Result<bool> {
+        let response = self.channel.send(IsRunning).await?;
 
-        response.await.unwrap()
+        Ok(response.await?)
     }
 
+    /// Start a node with the provided config, if it's not running
     pub async fn start(&self, config: WasmNodeConfig) -> Result<()> {
         let command = StartNode(config);
-        let response = self.channel.send(command);
+        let response = self.channel.send(command).await?;
 
-        Ok(response.await.unwrap()?)
+        Ok(response.await??)
     }
 
-    pub async fn local_peer_id(&self) -> String {
-        let response = self.channel.send(GetLocalPeerId);
+    /// Get node's local peer ID.
+    pub async fn local_peer_id(&self) -> Result<String> {
+        let response = self.channel.send(GetLocalPeerId).await?;
 
-        response.await.unwrap()
+        Ok(response.await?)
     }
 
+    /// Get current [`PeerTracker`] info.
     pub async fn peer_tracker_info(&self) -> Result<JsValue> {
-        let response = self.channel.send(GetPeerTrackerInfo);
+        let response = self.channel.send(GetPeerTrackerInfo).await?;
 
-        Ok(to_value(&response.await.unwrap())?)
+        Ok(to_value(&response.await?)?)
     }
 
-    pub async fn wait_connected(&self) {
+    /// Wait until the node is connected to at least 1 peer.
+    pub async fn wait_connected(&self) -> Result<()> {
         let command = WaitConnected { trusted: false };
-        let response = self.channel.send(command);
+        let response = self.channel.send(command).await?;
 
-        response.await.unwrap()
+        Ok(response.await?)
     }
-    pub async fn wait_connected_trusted(&self) {
+
+    /// Wait until the node is connected to at least 1 trusted peer.
+    pub async fn wait_connected_trusted(&self) -> Result<()> {
         let command = WaitConnected { trusted: false };
-        let response = self.channel.send(command);
+        let response = self.channel.send(command).await?;
 
-        response.await.unwrap()
+        Ok(response.await?)
     }
 
+    /// Get current network info.
     pub async fn network_info(&self) -> Result<NetworkInfoSnapshot> {
-        let response = self.channel.send(GetNetworkInfo);
+        let response = self.channel.send(GetNetworkInfo).await?;
 
-        Ok(response.await.unwrap())
+        Ok(response.await?)
     }
 
+    /// Get all the multiaddresses on which the node listens.
     pub async fn listeners(&self) -> Result<Array> {
-        let response = self.channel.send(GetListeners);
-        let response = response
-            .await
-            .unwrap()
-            .iter()
-            .map(js_value_from_display)
-            .collect();
+        let response = self.channel.send(GetListeners).await?;
+        let response = response.await?.iter().map(js_value_from_display).collect();
 
         Ok(response)
     }
 
+    /// Get all the peers that node is connected to.
     pub async fn connected_peers(&self) -> Result<Array> {
-        let response = self.channel.send(GetConnectedPeers);
-        let response = response
-            .await
-            .unwrap()
-            .iter()
-            .map(js_value_from_display)
-            .collect();
+        let response = self.channel.send(GetConnectedPeers).await?;
+        let response = response.await?.iter().map(js_value_from_display).collect();
 
         Ok(response)
     }
 
+    /// Trust or untrust the peer with a given ID.
     pub async fn set_peer_trust(&self, peer_id: &str, is_trusted: bool) -> Result<()> {
         let command = SetPeerTrust {
             peer_id: peer_id.to_string(),
             is_trusted,
         };
-        let response = self.channel.send(command);
+        let response = self.channel.send(command).await?;
 
-        Ok(response.await.unwrap())
+        Ok(response.await?)
     }
 
+    /// Request the head header from the network.
     pub async fn request_head_header(&self) -> Result<JsValue> {
         let command = RequestHeader(SingleHeaderQuery::Head);
-        let response = self.channel.send(command);
-        Ok(to_value(&response.await.unwrap()?)?)
+        let response = self.channel.send(command).await?;
+        Ok(to_value(&response.await??)?)
     }
 
+    /// Request a header for the block with a given hash from the network.
     pub async fn request_header_by_hash(&self, hash: &str) -> Result<JsValue> {
         let command = RequestHeader(SingleHeaderQuery::ByHash(hash.parse()?));
-        let response = self.channel.send(command);
+        let response = self.channel.send(command).await?;
 
-        Ok(to_value(&response.await.unwrap()?)?)
+        Ok(to_value(&response.await??)?)
     }
 
+    /// Request a header for the block with a given height from the network.
     pub async fn request_header_by_height(&self, height: u64) -> Result<JsValue> {
         let command = RequestHeader(SingleHeaderQuery::ByHeight(height));
-        let response = self.channel.send(command);
+        let response = self.channel.send(command).await?;
 
-        Ok(to_value(&response.await.unwrap()?)?)
+        Ok(to_value(&response.await??)?)
     }
 
+    /// Request headers in range (from, from + amount] from the network.
+    ///
+    /// The headers will be verified with the `from` header.
     pub async fn request_verified_headers(
         &self,
         from_header: JsValue,
@@ -177,7 +188,7 @@ impl NodeDriver {
     ) -> Result<Array> {
         let from = from_value(from_header)?;
         let command = RequestMultipleHeaders(MultipleHeaderQuery::GetVerified { from, amount });
-        let response = self.channel.send(command);
+        let response = self.channel.send(command).await?;
 
         let result = response
             .await
@@ -188,200 +199,43 @@ impl NodeDriver {
         Ok(result)
     }
 
-    pub async fn syncer_info(&self) -> Result<JsValue> {
-        let response = self.channel.send(GetSyncerInfo);
-
-        Ok(to_value(&response.await.unwrap())?)
-    }
-
-    pub async fn get_network_head_header(&self) -> Result<JsValue> {
-        let command = LastSeenNetworkHead;
-        let response = self.channel.send(command);
-
-        Ok(to_value(&response.await.unwrap())?)
-    }
-
-    pub async fn get_local_head_header(&self) -> Result<JsValue> {
-        let command = GetHeader(SingleHeaderQuery::Head);
-        let response = self.channel.send(command);
-
-        Ok(to_value(&response.await.unwrap())?)
-    }
-
-    pub async fn get_header_by_hash(&self, hash: &str) -> Result<JsValue> {
-        let command = GetHeader(SingleHeaderQuery::ByHash(hash.parse()?));
-        let response = self.channel.send(command);
-
-        Ok(to_value(&response.await.unwrap())?)
-    }
-    pub async fn get_header_by_height(&self, height: u64) -> Result<JsValue> {
-        let command = GetHeader(SingleHeaderQuery::ByHeight(height));
-        let response = self.channel.send(command);
-
-        Ok(to_value(&response.await.unwrap())?)
-    }
-    pub async fn get_headers(
-        &self,
-        start_height: Option<u64>,
-        end_height: Option<u64>,
-    ) -> Result<Array> {
-        let command = GetMultipleHeaders(MultipleHeaderQuery::Range {
-            start_height,
-            end_height,
-        });
-        let response = self.channel.send(command);
-        let result = response
-            .await
-            .iter()
-            .map(|h| to_value(&h).unwrap())
-            .collect();
-
-        Ok(result)
-    }
-    pub async fn get_sampling_metadata(&self, height: u64) -> Result<JsValue> {
-        let command = GetSamplingMetadata { height };
-        let response = self.channel.send(command);
-
-        Ok(to_value(&response.await.unwrap())?)
-    }
-}
-
-/*
-#[wasm_bindgen(js_class = Node)]
-impl WasmNode {
-    /// Create a new Lumina node.
-    #[wasm_bindgen(constructor)]
-    pub async fn new(config: WasmNodeConfig) -> Result<WasmNode> {
-        let config = config.into_node_config().await?;
-
-        if let Ok(store_height) = config.store.head_height().await {
-            info!("Initialised store with head height: {store_height}");
-        } else {
-            info!("Initialized new empty store");
-        }
-
-        let node = Node::new(config)
-            .await
-            .js_context("Failed to start the node")?;
-
-        Ok(Self(node))
-    }
-
-    /// Get node's local peer ID.
-    pub fn local_peer_id(&self) -> String {
-        self.0.local_peer_id().to_string()
-    }
-
-    /// Get current [`PeerTracker`] info.
-    pub fn peer_tracker_info(&self) -> Result<JsValue> {
-        let peer_tracker_info = self.0.peer_tracker_info();
-        Ok(to_value(&peer_tracker_info)?)
-    }
-
-    /// Wait until the node is connected to at least 1 peer.
-    pub async fn wait_connected(&self) -> Result<()> {
-        Ok(self.0.wait_connected().await?)
-    }
-
-    /// Wait until the node is connected to at least 1 trusted peer.
-    pub async fn wait_connected_trusted(&self) -> Result<()> {
-        Ok(self.0.wait_connected_trusted().await?)
-    }
-
-    /// Get current network info.
-    pub async fn network_info(&self) -> Result<NetworkInfoSnapshot> {
-        Ok(self.0.network_info().await?.into())
-    }
-
-    /// Get all the multiaddresses on which the node listens.
-    pub async fn listeners(&self) -> Result<Array> {
-        let listeners = self.0.listeners().await?;
-
-        Ok(listeners
-            .iter()
-            .map(js_value_from_display)
-            .collect::<Array>())
-    }
-
-    /// Get all the peers that node is connected to.
-    pub async fn connected_peers(&self) -> Result<Array> {
-        Ok(self
-            .0
-            .connected_peers()
-            .await?
-            .iter()
-            .map(js_value_from_display)
-            .collect::<Array>())
-    }
-
-    /// Trust or untrust the peer with a given ID.
-    pub async fn set_peer_trust(&self, peer_id: &str, is_trusted: bool) -> Result<()> {
-        let peer_id = peer_id.parse().js_context("Parsing peer id failed")?;
-        Ok(self.0.set_peer_trust(peer_id, is_trusted).await?)
-    }
-
-    /// Request the head header from the network.
-    pub async fn request_head_header(&self) -> Result<JsValue> {
-        let eh = self.0.request_head_header().await?;
-        Ok(to_value(&eh)?)
-    }
-
-    /// Request a header for the block with a given hash from the network.
-    pub async fn request_header_by_hash(&self, hash: &str) -> Result<JsValue> {
-        let hash: Hash = hash.parse()?;
-        let eh = self.0.request_header_by_hash(&hash).await?;
-        Ok(to_value(&eh)?)
-    }
-
-    /// Request a header for the block with a given height from the network.
-    pub async fn request_header_by_height(&self, height: u64) -> Result<JsValue> {
-        let eh = self.0.request_header_by_height(height).await?;
-        Ok(to_value(&eh)?)
-    }
-
-    /// Request headers in range (from, from + amount] from the network.
-    ///
-    /// The headers will be verified with the `from` header.
-    pub async fn request_verified_headers(&self, from: JsValue, amount: u64) -> Result<Array> {
-        let header =
-            from_value::<ExtendedHeader>(from).js_context("Parsing extended header failed")?;
-        let verified_headers = self.0.request_verified_headers(&header, amount).await?;
-
-        Ok(verified_headers
-            .iter()
-            .map(to_value)
-            .collect::<StdResult<_, _>>()?)
-    }
-
     /// Get current header syncing info.
     pub async fn syncer_info(&self) -> Result<JsValue> {
-        let syncer_info = self.0.syncer_info().await?;
-        Ok(to_value(&syncer_info)?)
+        let response = self.channel.send(GetSyncerInfo).await?;
+
+        Ok(to_value(&response.await?)?)
     }
 
     /// Get the latest header announced in the network.
-    pub fn get_network_head_header(&self) -> Result<JsValue> {
-        let maybe_head_header = self.0.get_network_head_header();
-        Ok(to_value(&maybe_head_header)?)
+    pub async fn get_network_head_header(&self) -> Result<JsValue> {
+        let command = LastSeenNetworkHead;
+        let response = self.channel.send(command).await?;
+
+        Ok(to_value(&response.await?)?)
     }
 
     /// Get the latest locally synced header.
     pub async fn get_local_head_header(&self) -> Result<JsValue> {
-        let local_head = self.0.get_local_head_header().await?;
-        Ok(to_value(&local_head)?)
+        let command = GetHeader(SingleHeaderQuery::Head);
+        let response = self.channel.send(command).await?;
+
+        Ok(to_value(&response.await?)?)
     }
 
     /// Get a synced header for the block with a given hash.
     pub async fn get_header_by_hash(&self, hash: &str) -> Result<JsValue> {
-        let hash: Hash = hash.parse().js_context("parsing hash failed")?;
-        let eh = self.0.get_header_by_hash(&hash).await?;
-        Ok(to_value(&eh)?)
+        let command = GetHeader(SingleHeaderQuery::ByHash(hash.parse()?));
+        let response = self.channel.send(command).await?;
+
+        Ok(to_value(&response.await?)?)
     }
 
     /// Get a synced header for the block with a given height.
     pub async fn get_header_by_height(&self, height: u64) -> Result<JsValue> {
-        let eh = self.0.get_header_by_height(height).await?;
-        Ok(to_value(&eh)?)
+        let command = GetHeader(SingleHeaderQuery::ByHeight(height));
+        let response = self.channel.send(command).await?;
+
+        Ok(to_value(&response.await?)?)
     }
 
     /// Get synced headers from the given heights range.
@@ -397,40 +251,29 @@ impl WasmNode {
         &self,
         start_height: Option<u64>,
         end_height: Option<u64>,
-    ) -> Result<JsValue> {
-        let headers = match (start_height, end_height) {
-            (None, None) => self.0.get_headers(..).await,
-            (Some(start), None) => self.0.get_headers(start..).await,
-            (None, Some(end)) => self.0.get_headers(..=end).await,
-            (Some(start), Some(end)) => self.0.get_headers(start..=end).await,
-        }?;
+    ) -> Result<Array> {
+        let command = GetMultipleHeaders(MultipleHeaderQuery::Range {
+            start_height,
+            end_height,
+        });
+        let response = self.channel.send(command).await?;
+        let result = response
+            .await
+            .iter()
+            .map(|h| to_value(&h).unwrap())
+            .collect();
 
-        Ok(to_value(&headers)?)
+        Ok(result)
     }
 
     /// Get data sampling metadata of an already sampled height.
     pub async fn get_sampling_metadata(&self, height: u64) -> Result<JsValue> {
-        let metadata = self.0.get_sampling_metadata(height).await?;
+        let command = GetSamplingMetadata { height };
+        let response = self.channel.send(command).await?;
 
-        #[derive(Serialize)]
-        struct Intermediate {
-            accepted: bool,
-            cids_sampled: Vec<String>,
-        }
-
-        let metadata = metadata.map(|m| Intermediate {
-            accepted: m.accepted,
-            cids_sampled: m
-                .cids_sampled
-                .into_iter()
-                .map(|cid| cid.to_string())
-                .collect(),
-        });
-
-        Ok(to_value(&metadata)?)
+        Ok(to_value(&response.await?)?)
     }
 }
-*/
 
 #[wasm_bindgen(js_class = NodeConfig)]
 impl WasmNodeConfig {
